docker_inf_data/requirements/requirements_iskandre_inf_pip.txt
absl-py==1.4.0
accelerate==0.18.0
addict==2.4.0
aenum==3.1.12
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal #@ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1667935791922/work
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
astunparse==1.6.3
async-timeout #@ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1640026696943/work
attrs #@ file:///home/conda/feedstock_root/build_artifacts/attrs_1683424013410/work
basicsr==1.4.2
blendmodes==2022
blinker #@ file:///home/conda/feedstock_root/build_artifacts/blinker_1681349778161/work
boltons==23.0.0
brotlipy #@ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1648854167867/work
cachetools #@ file:///home/conda/feedstock_root/build_artifacts/cachetools_1674482203741/work
certifi==2023.5.7
cffi #@ file:///home/conda/feedstock_root/build_artifacts/cffi_1636046050867/work
chardet==5.1.0
charset-normalizer==3.1.0
clean-fid==0.1.35
click #@ file:///home/conda/feedstock_root/build_artifacts/click_1666798198223/work
clip-anytorch==2.5.2
cmake==3.26.3
contourpy==1.0.7
cryptography #@ file:///home/conda/feedstock_root/build_artifacts/cryptography_1637687018854/work
cycler==0.11.0
decorator #@ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
deprecation==2.1.0
docker-pycreds==0.4.0
einops==0.6.1
entrypoints==0.4
facexlib==0.3.0
fastapi==0.95.1
ffmpy==0.3.0
filelock==3.12.0
filterpy==1.4.5
font-roboto==0.0.1
fonts==0.0.3
fonttools==4.39.3
frozenlist #@ file:///croot/frozenlist_1670004507010/work
fs==2.4.16
fs-gcsfs==1.5.1
fsspec #@ file:///home/conda/feedstock_root/build_artifacts/fsspec_1683494881189/work
ftfy==6.1.1
future==0.18.3
gcsfs #@ file:///home/conda/feedstock_root/build_artifacts/gcsfs_1683503493718/work
gfpgan==1.3.8
gitdb==4.0.10
GitPython==3.1.31
google-api-core #@ file:///home/conda/feedstock_root/build_artifacts/google-api-core-split_1670580271439/work
google-auth #@ file:///home/conda/feedstock_root/build_artifacts/google-auth_1681369121356/work
google-auth-oauthlib #@ file:///home/conda/feedstock_root/build_artifacts/google-auth-oauthlib_1675825735130/work
google-cloud-core #@ file:///home/conda/feedstock_root/build_artifacts/google-cloud-core_1658759191372/work
google-cloud-storage #@ file:///home/conda/feedstock_root/build_artifacts/google-cloud-storage_1683272316756/work
google-crc32c #@ file:///home/conda/feedstock_root/build_artifacts/google-crc32c_1651517205891/work
google-resumable-media #@ file:///home/conda/feedstock_root/build_artifacts/google-resumable-media_1682369385616/work
googleapis-common-protos #@ file:///home/conda/feedstock_root/build_artifacts/googleapis-common-protos-feedstock_1672913962683/work
gradio==3.28.1
gradio_client==0.1.4
grpcio==1.54.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.14.1
idna #@ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
imageio==2.28.1
inflection==0.5.1
Jinja2==3.1.2
jsonmerge==1.9.0
jsonschema==4.17.3
k-diffusion==0.0.14
kiwisolver==1.4.4
kornia==0.6.12
lark==1.1.5
lazy_loader==0.2
linkify-it-py==2.0.2
lit==16.0.2
llvmlite==0.40.0
lmdb==1.4.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numba==0.57.0
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib #@ file:///home/conda/feedstock_root/build_artifacts/oauthlib_1666056362788/work
omegaconf==2.3.0
open-clip-torch==2.19.0
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
orjson==3.8.11
packaging==23.1
pandas==2.0.1
pathtools==0.1.2
piexif==1.1.3
Pillow==9.5.0
protobuf==3.20.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser #@ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydantic==1.10.7
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.15.1
PyJWT #@ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1666240235902/work
pyOpenSSL #@ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing==3.0.9
pyre-extensions==0.0.29
pyrsistent==0.19.3
PySocks #@ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==1.7.7
pytz==2023.3
pyu2f #@ file:///home/conda/feedstock_root/build_artifacts/pyu2f_1604248910016/work
PyWavelets==1.4.1
PyYAML==6.0
realesrgan==0.3.0
regex==2023.5.5
requests #@ file:///home/conda/feedstock_root/build_artifacts/requests_1682535435083/work
requests-oauthlib #@ file:///home/conda/feedstock_root/build_artifacts/requests-oauthlib_1643557462909/work
resize-right==0.0.2
rich==13.3.5
rsa #@ file:///home/conda/feedstock_root/build_artifacts/rsa_1658328885051/work
safetensors==0.3.1
scikit-image==0.20.0
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.21.1
setproctitle==1.3.2
six #@ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.0
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
tb-nightly==2.13.0a20230503
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tifffile==2023.4.12
timm==0.4.12
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
torchdiffeq==0.2.3
torchmetrics==0.11.4
torchsde==0.2.5
torchvision==0.15.1
tqdm==4.65.0
trampoline==0.1.2
transformers==4.25.1
triton==2.0.0
typing-inspect==0.8.0
typing_extensions #@ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1678559861143/work
tzdata==2023.3
uc-micro-py==1.0.2
urllib3 #@ file:///home/conda/feedstock_root/build_artifacts/urllib3_1678635778344/work
uvicorn==0.22.0
wandb==0.15.1
wcwidth==0.2.6
websockets==11.0.2
Werkzeug==2.3.3
xformers==0.0.19
yapf==0.33.0
yarl==1.9.2
